CLAVIN/requirements.txt
argparse==1.2.1
beautifulsoup4==4.3.2
configparser==3.3.0r2
oauthlib==0.6.3
pymongo==2.7.2
python-dateutil==2.2
requests==2.3.0
requests-oauthlib==0.4.1